requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.9.13

# Configuration
django-environ==0.4.0

# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.26.1


# Python-PostgreSQL and MySQL Database Adapters
psycopg2==2.6.2
mysqlclient

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.4
redis>=2.10.0

# Celery
celery==3.1.23

django_compressor==2.0

# Admin
django-suit

# Your custom requirements go here
tqdm
git+http://github.com/javipalanca/django-wordpress.git
django-reversion
django-wysiwyg-redactor
django-suit-redactor
beautifulsoup4
django-disqus
django-social-share
django-categories
django-taggit
django-taggit-autosuggest
git+http://github.com/javipalanca/django-admin-bootstrapped.git
django-admin-bootstrapped-plus
django-bootstrap3
django-guardian==1.4.5
django-adminactions
django-likert-field
django-phonenumber-field
geopy
django-leaflet
django-envelope
dj-database-url
google-api-python-client
django-analytical
django-request
django-hitcount
django-newsletter
requests-facebook
tweepy
django-haystack
elasticsearch>=2.0.0,<3.0.0

# django rest framework
djangorestframework
djangorestframework-gis
django-taggit-serializer
drf-haystack

django-maintenance-mode


# Amazon Simple Email Service SES
django-ses
